web: remove dead ends
We want to use the tacd web interface as a demonstration at our booth
at an industrial fair with only a touchscreen and a browser without
navigation buttons as interface.
This means we can not have any links that do not lead back to the
web interface.

This should not be merged into the main branch, as this is not a normal
requirement for us.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>