regulators: handle iobus and vuart power via regulator instead of gpio
We have previously used GPIOs to control the IOBus power supply on the TAC.
With Gen 2 of the LXA TAC hardware the polarity of that line has changed.
We now have the choice of either special-casing the respective GPIO in the
tacd or store the difference in the device tree.

I prefer the second option and have set up a regulator for the IOBus line
in the devicetree and exposed it via to userspace via userspace-consumer.c.
This has the added benefit of allowing us to control the 3.3V pin on the
DUT UART box connector as well (which sits on a dedicated LDO on the PMIC).

Use the userspace-consumer interface in the tacd.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>