screens: usb: Add an indication for the total current limit to the LCD
When there are e.g. three ports drawing 300mA each the information on the
display would previously look like everything was fine.
This is however not the case as 3 * 300mA is over the 700mA current limit,
meaning the ports would all run with reduced output voltage.
Add an indicator for the total current draw to help troubleshoot overcurrent
issues.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>