iobus: use output voltage and current to detect a fault
The fault pin from the smart fuse on board of the LXA TAC will rarely
(if ever) be active, as it only indicates a shutdown due to the smart
fuse overheating.

If the fuse is sufficiently cooled the IOBus may be in current limiting
condition indefinitly without the fault signal activating.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>