dut_power: make GPIOs open drain
This improves EMI performance of the TAC a tiny bit, by not driving the
pins on the flat flex using the somewhat noisy VCC_IO power supply.
Signals on the flat flex are prone to coupling to one another due to
missing GND reference plane.
We _think_ we see this kind of noise on the IOBus lines, which are are
present on the flat-flex, as well as on the external IOBus connector.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>