HACK: dut_power: completely remove the inverted polarity check
There are certain combinations of DUT and TAC power supplies and
how connections between DUT and TAC are set up, that result in "ghost"
voltage measurements on the TACs power board.

To our current knowledge these are not measurement errors but instead
actual voltages on the DUT power supply, that can also be verified
using an oscilloscope or a multimeter in AC mode.

We are looking for a solution for this issue that does not involve
just reducing the input impedance of the measurement (which would
likely solve it, but would come with other drawbacks).

Provide a hack that works around the problem in software while we
come up with an actual solution.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>